+++ tests/testthat/test-loading_workbook.R
@@ -331,15 +331,3 @@ test_that("vml target is updated on load", {
   expect_equal(exp, got)
 
 })
-
-test_that("load file with x namespace", {
-
-  fl <- "https://github.com/ycphs/openxlsx/files/8480120/2022-04-12-11-42-36-DP_Melanges1.xlsx"
-
-  expect_warning(
-    wb <- wb_load(fl),
-    "This file was slightly modified to import it. It should be fine, but be careful."
-  )
-  expect_false(getOption("openxlsx2.namespace_x", default = FALSE))
-
-})