+++ tests/testthat/test-read_sources.R
@@ -191,3 +191,15 @@ test_that("reading charts", {
   expect_equal("", wb$drawings_rels[[21]])
 
 })
+
+test_that("load file with xml namespace", {
+
+  fl <- "https://github.com/ycphs/openxlsx/files/8480120/2022-04-12-11-42-36-DP_Melanges1.xlsx"
+
+  expect_warning(
+    wb <- wb_load(fl),
+    "has been removed from the xml files, for example"
+  )
+  expect_null(getOption("openxlsx2.namespace_xml"))
+
+})